+++ R/mock_edish.R
@@ -7,7 +7,7 @@ mock_edish_app <- function() {
   dm <- pharmaverseadam::adsl |> dplyr::mutate(dplyr::across(dplyr::where(is.character), as.factor))
   lb <- pharmaverseadam::adlb |> dplyr::mutate(dplyr::across(dplyr::where(is.character), as.factor))
 
-  mock_edish_UI <- function() { # nolint
+-  mock_edish_UI <- function() { # nolint
     shiny::fluidPage(edish_UI(
       module_id = "edish",
       arm_default_vals = "Xanomeline High Dose",
@@ -45,6 +45,9 @@ mock_edish_app <- function() {
 mock_edish_mm <- function() {
   dm <- pharmaverseadam::adsl
   lb <- pharmaverseadam::adlb
+
+  attr(dm, "meta") <- base::file.info("NEWS.md")
+  attr(lb, "meta") <- base::file.info("NEWS.md")
 
   module_list <- list(
     "eDISH Demo" = mod_edish(