+++ NEWS.md
@@ -1,6 +1,7 @@
 # dv.edish 2.0.2-9002
 
 - Implement table of subject counts and percentages in areas delimited by the specified reference lines.
+- Ensure reference line values appear on the axes.
 
 # dv.edish 2.0.2-9001